feat: Implementar funcionalidad del bot para la gestión de archivos y recordatorios de entrega - Se añadió el cog `Archivos` para subir, ver, eliminar y enviar archivos. - Se introdujo el cog `Entregas` para gestionar plazos de entrega con recordatorios. - Se creó el cog `Otros` para comandos generales como saludos y ayuda.- Se estableció el esquema de base de datos para archivos y entregas. - Se implementaron funciones utilitarias para el acceso seguro a diccionarios.- Se inicializó el bot con los intents necesarios y la configuración de la base de datos. - Se actualizaron los requisitos para incluir los paquetes necesarios.
