+++ bot/__init__.py
@@ -1,0 +1,1 @@
+# Inicialización del paquete bot
