+++ bot/utils.py
@@ -1,0 +1,7 @@
+# Utilidades generales para el bot
+
+def safe_get(d, key, default=None):
+    try:
+        return d[key]
+    except (KeyError, TypeError):
+        return default
