+++ requirements.txt
@@ -1,0 +1,3 @@
+discord.py>=2.3.0
+aiofiles
+requests
